move KeyBackedConfig to a separate file
